Refactor cli and break tests
