--- cli/mock/test.json
@@ -1,3 +1,0 @@
-{
-    "hello": "world"
-}
